requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.6
decorator==5.1.1
Django==5.1.3
django-extensions==3.2.3
django-modeltranslation==0.19.12
djangorestframework==3.15.2
executing==2.1.0
frozenlist==1.5.0
gTTS==2.5.3
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
magic-filter==1.0.12
matplotlib-inline==0.1.7
multidict==6.1.0
multipledispatch==1.0.0
nest-asyncio==1.6.0
ngrok==1.4.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.1
paramiko==3.5.0
parso==0.8.4
pefile==2024.8.26
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.0
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMySQL==1.1.1
PyNaCl==1.5.0
PyOpenGL==3.1.7
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32==307
pywin32-ctypes==0.2.3
pyzmq==26.2.0
requests==2.32.3
setuptools==75.8.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
unitils==0.1.2
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.17.1
